refactor bloom model
